requirements.txt
streamlit
pandas
python-docx
pdfplumber
openai